minor adjustment
